# internal/web/docs/openapi.yaml
openapi: 3.1.0
info:
  version: 0.0.1
  title: Leagueify API
  summary: Open Source Sporting League Platform.
  description: Host your own sports league using Leagueify, the all-in-one league hosting platform.
  license:
    name: MIT
    url: https://raw.githubusercontent.com/Leagueify/api/main/LICENSE
  contact:
    name: Leagueify Support
    url: https://github.com/Leagueify/api/issues
servers:
  - url: http://localhost/api
    description: Local Development Server - Docker
  - url: http://localhost:8888/api
    description: Local Development Server - Standalone Executable
paths:
...
  /players:
    get:
      tags:
        - Players
      summary: Get All Players
      description: '
        Get all players for the active account
        '
      security:
        - apiKey: []
      responses:
        200:
          description: Registered Player IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    description: List of player IDs on the active account
                    type: array
                    items:
                      description: Player ID
                      type: string
              examples:
                accountPlayers:
                  summary: Player IDs on active account
                  value: {
                    "players": [
                        "6G37TEN",
                        "123ABCD"
                      ]
                    }
        401:
          $ref: "#/components/errors/unauthorized"
    post:
      tags:
        - Players
      summary: Create Player(s)
      description: '
        Create one or more players for the active account
        '
      security:
        - apiKey: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                players:
                  description: List of Players to create
                  type: array
                  items:
                    $ref: "#/components/players/schema"
      responses:
        201:
          description: Player(s) Created
          content:
            application/json:
              schema:
                $ref: "#/components/successful/schema"
              examples:
                playersCreated:
                  $ref: "#/components/successful/example"
        400:
          $ref: "#/components/errors/badRequest"
        401:
          $ref: "#/components/errors/unauthorized"
  /players/{id}:
    delete:
      tags:
        - Players
      summary: Delete Player
      description: '
        Delete a player from the active account
        '
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: ID of the player to delete
          required: true
          type: string
      responses:
        204:
          description: Player(s) Deleted
        400:
          $ref: "#/components/errors/badRequest"
        401:
          $ref: "#/components/errors/unauthorized"
    get:
      tags:
        - Players
      summary: Get Player Information
      description: '
        Retrieve player information
        '
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: ID of the player to retrieve
          required: true
          type: string
      responses:
        200:
          description: Player Information
          content:
            application/json:
              schema:
                $ref: "#/components/players/schema"
        401:
          $ref: "#/components/errors/unauthorized"
  /players/register:
    post:
      tags:
        - Players
      summary: Register Player(s)
      description: '
        Register players within the league
        '
      security:
        - apiKey: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                players:
                  description: List of Players to delete
                  type: array
                  items:
                    description: Player ID
                    type: string
            examples:
              validPlayerDeleteRequest:
                summary: Delete request
                value: {
                  "players": [
                      "6G37TEN",
                      "123ABCD"
                    ]
                  }
      responses:
        201:
          description: Player(s) Created
          content:
            application/json:
              schema:
                $ref: "#/components/successful/schema"
              examples:
                playersCreated:
                  $ref: "#/components/successful/example"
        400:
          $ref: "#/components/errors/badRequest"
        401:
          $ref: "#/components/errors/unauthorized"
...
components:
  errors:
    badRequest:
      description: Error with request body
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: Error message detailing the request issue
                type: string
          examples:
            default:
              summary: Response containing error message
              value: {
                "detail": "< details about the error >",
                "status": "bad request"
                }
    unauthorized:
      description: Authorization Issue
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: Error message detailing the request issue
                type: string
          examples:
            default:
              summary: Response containing unauthorized message
              value: {
                "status": "unauthorized"
                }
  players:
    schema:
      type: object
      properties:
        firstName:
          description: Player's first name
          type: string
          minLength: 3
          maxLength: 32
          example: "Michael"
        lastName:
          description: Player's last name
          type: string
          minLength: 3
          maxLength: 32
          example: "duBois"
        dateOfBirth:
          description: Player's date of birth
          type: string
          example: "1990-08-31"
        position:
          description: Player's desired position
          type: string
          example: "goalie"
      required:
        - firstName
        - lastName
        - dateOfBirth
        - position
...
  successful:
    schema:
      type: object
      properties:
        status:
          description: Successful Request Response
          type: string
    example:
      summary: Successful Request Response
      value: {
          "status": "successful"
        }
  securitySchemes:
    apiKey:
      type: apiKey
      name: apiKey
      in: header